\name{refund.wave-internal}
\alias{checkError}
\alias{decorrelate}
\alias{decomp}
\alias{decomp2d}
\alias{decomp3d}
\alias{getRsq}
\alias{reconstr}
\alias{reconstr2d}
\alias{reconstr3d}
\alias{waveletGetResult}
\alias{waveletSetup}
\alias{waveletGetCV}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions for the refund.wave package
}
\description{
These functions are ordinarily not to be called by the user, but if you contact the package authors with any questions about them, we'll do our best to clarify matters.}


