\name{refund.wave-package}
\Rdversion{1.1}
\alias{refund.wave-package}
\alias{refund.wave}
\docType{package}
\title{
Wavelet-Domain Regression with Functional Data
}
\description{
Methods for regression with functional data in the wavelet domain. Various approaches to regression with scalar responses and functional predictors are implemented by \code{\link{wcr}} and \code{\link{wnet}}.}
\details{For a complete list of functions type \code{library(help=refund.wave)}.
}

\author{
Lan Huo \email{lan.huo@nyumc.org}, Philip Reiss \email{phil.reiss@nyumc.org} and 
Yihong Zhao \email{yihong.zhao@nyumc.org}
}
%\keyword{ package }

